+++ env/dev/config/constants-local.php
@@ -1,0 +1,4 @@
+<?php
+
+defined('APP_ENV') or define('APP_ENV', 'dev');
+defined('APP_DEBUG') or define('APP_DEBUG', true);